+++ HISTORY.rst
@@ -10,3 +10,18 @@ History
 0.2.1 (2020-02-28)
 ------------------
 * add support to download files and archives from the web and google drive
+
+0.3.0 (2020-10-04)
+-------------------
+
+* addedd support for removing categories and other utilities
+
+0.4.0 (2020-10-05)
+-------------------
+
+* addedd support to create a dataset from scratch
+
+0.5.0 (2020-10-06)
+-------------------
+
+* addedd support to visualize images and annotations